first real commit index.html
Hello to who ever is seeing this. So, I want to clear some stuff up. I was talking to my friends who are participating at this hackathon at hack club and found out about what github commits are. I did not know what they were up until now and did not know that I should be regularly doing them. So, this is my first actual commit where I showed what I worked on. Up until now I would only upload to github when I wanted to host a website or when I was submitting a project. I am sorry for any inconvenience this could have caused. If anybody has any questions on how I got to this point and how I coded this since it may not be clear with the commits, feel free to reach out to me, I would be happy to answer your questions. 

So, this is the first version of my portfolio website, I am still working on it and adding more features now.

Thanks for understanding.