Update index.html
added extra features and spots for new links. Been working on this for the past 3 hours. new animaitons ad hover effects, fixed content with new updates. Almost done with css for this. Done with the js. Still gotta fix the image links, the links to my socials, and some content, have to fix some formatting issues as well.